+++ lib/typescript/commonjs/index.d.ts
@@ -7,7 +7,7 @@ export { clearAllCollections, devClearAllDataAndState, resetAllModelsState } fro
 export { stableSerialize } from './core/serialize';
 export { createExtractSink, createMutationExtractResolver, liftExtractNodes } from './core/extract';
 export { invalidateDbRequests, invalidateModel, resetDbQueryRuntime } from './core/queryClient';
-export { createDbSubscriptionRuntime } from './core/subscriptionRuntime';
+export { createDbSubscriptionRuntime, defineDbSubscriptionEntry } from './core/subscriptionRuntime';
 export { patchWhenPresent, waitForRow } from './core/rowWaiters';
 export { f } from './schema/f';
 export { compositeId } from './schema/schema';